### 2023-09-07 #105 @lucaferranti
Hi @valvarezapa :wave: ,

Apologies for the long delay, I'll be the editor for this submission.

Since it has been a while, I firsted wanted to double check with you, **are you still interested in having this published?**

If yes, I will start looking for potential reviewers. Suggestions are also welcome of course.

As a first step, as you can see from [this comment](https://github.com/JuliaCon/proceedings-review/issues/105#issuecomment-1710611310) currently the paper fails to compile. This should be fixed before starting the review process.

Looking at the error message, I suspect in your case the error is because you are using the `minted` library (for formatting code snippets I presume) but the compiler cannot find the library file. This is because if you use extra packages with `\usepackage` you should also include the corresponding source code into the `paper` folder, since we currently the bot cannot download those automatically. Can you look into this?

Please let me know if you have any further questions
