### 2023-09-08 b85b70a
Deleted minted from paper.tex
