### 2023-09-08 #105 @valvarezapa
Hi!

Yes! I am still interested in having the paper published. I am very glad to
hear about this.

I have just modified the *paper.tex* file so it is now self-contained and
does not depend on any extra packages. I hope the bot can compile it now.

If there is anything else I can do to make the publishing process easier,
please let me know. Thank you for your time.
Best regards,
Víctor

<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
Libre
de virus.www.avast.com
<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
<#DAB4FAD8-2DD7-40BB-A1B8-4E2AA1F9FDF2>

El jue, 7 sept 2023 a las 20:49, Luca Ferranti ***@***.***>)
escribió:

> Hi @valvarezapa <https://github.com/valvarezapa> 👋 ,
>
> Apologies for the long delay, I'll be the editor for this submission.
>
> Since it has been a while, I firsted wanted to double check with you, *are
> you still interested in having this published?*
>
> If yes, I will start looking for potential reviewers. Suggestions are also
> welcome of course.
>
> As a first step, as you can see from this comment
> <https://github.com/JuliaCon/proceedings-review/issues/105#issuecomment-1710611310>
> currently the paper fails to compile. This should be fixed before starting
> the review process.
>
> Looking at the error message, I suspect in your case the error is because
> you are using the minted library (for formatting code snippets I presume)
> but the compiler cannot find the library file. This is because if you use
> extra packages with \usepackage you should also include the corresponding
> source code into the paper folder, since we currently the bot cannot
> download those automatically. Can you look into this?
>
> Please let me know if you have any further questions
>
