### 2023-10-17 #105 @valvarezapa
Thank you very much, Luca!

<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
Libre
de virus.www.avast.com
<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
<#DAB4FAD8-2DD7-40BB-A1B8-4E2AA1F9FDF2>

El mar, 17 oct 2023 a las 9:16, Luca Ferranti ***@***.***>)
escribió:

> Hi @valvarezapa <https://github.com/valvarezapa>, quick update from my
> side.
>
> I found a first reviewer. I'll try to find a second one asap.
>
> @lbenet <https://github.com/lbenet> thank you for volunteering as
> reviewer. You can find more details on the juliacon review process here
> <https://juliacon.github.io/proceedings-guide/reviewer/>. Note this
> submission is still in pre-review and the review will start once I found a
> second reviewer. Meanwhile, feel free to ask me any questions you have.
>
