### 2024-02-13 #140 @whedon
[REVIEW]: Lyapunov Cycle Detector (LCD.jl): A new tool for the study of basins of attraction induced by rational maps
**Submitting author:** @valvarezapa (<a href="http://orcid.org/0000-0001-8232-1809">Víctor Álvarez</a>)
**Repository:** <a href="https://github.com/valvarezapa/LCD" target ="_blank">https://github.com/valvarezapa/LCD</a>
**Version:**

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [ ] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
