### 2024-02-28 #140 @lbenet
First, apologies for the delay to send this review.

The paper is a readable and concise extended abstract, describing the authors package, which is rather specilized in the study of the basins of attraction of rational complex maps, without a priori requiring the computation of the fixed-points. I find the paper clear, it includes references which may address some of the checklist points that I left unchecked, such as the details and proofs of the methodology. While I am in favor of an eventual acceptance, I think the repo and the paper need some improvements to make the life easier to any potential user of the package; some concrete points are listed below. This is the reason that I recommend to make minor modifications in the paper and the repo.

- Very important: there is no `Project.toml` file, so it is not obvious what dependencies are required (`Polynomials`, `PyPlot` and `Colors`); this point must be addressed by the author, since it complicates installing the package.

- The instructions to install the package (in the `README.md` file) are outdated and do not work. The author should include a clause in the paper stating that the package is not in the General Registry, and include how to install it. These instructions should also be included in the `README.md` file, as well as in the `LCD Quick User Guide.ipynb` file. (I managed to clone locally the repo, and after installing its dependencies one by one I could load the package with `include`; the fact that there is no `Project.toml` stops Julia from installing it.)

- There is no documentation of the package as such, which for this type of submission may be excused. Yet, there is a Jupyter notebook, `LCD Quick User Guide.ipynb`, which is very useful but outdated, and among other is aimed to illustrate many aspects of the package usage. I suggest to add a link to that file in the `README.md` file, and also update it as needed, so it can be used. Among other, that file clarifies what `coefnum` and `coefden` represent in the paper, and how to constructed them; in the paper these quantities are only mentioned with no details.

- Along the same lines, the author should add in the paper, perhaps as the figure captions of the figures included, the coefficients used to produce those figures, which incidentally are not mentioned in the paper. Importantly, I could not produce the image of the basin of attraction for $z^3-1$ using the command in the paper; I only obtained a uniform grey box; the same occurs following the LCD Quick User Guide. This may be due to a limitation of something I didn't install, which can be solved by making explicitly what is truly needed.

- The `README.md` file in the repo, which has lots of details, includes no references and seems outdated; this should be easily fixed. There, there are several mentions to the preprint "Mathematical Framework of LCD" (which I guess is Ref [2] in the paper), and the reader is invited to check it for details, without a link nor a proper reference. Since this file has not been modified for two years, I strongly suggest to update it, and consider updating Reference [2] in the paper, which is hopefully published by now.

- There are no tests for the package, which again might be excused; yet, I strongly recommend to include at least some. Easy are the fixed points simple complex polynomials. More elaborated tests certainly be appreciated.

- Regarding the paper itself, there is a typo: it is written "forming this calculations" and should be "forming these calculations". Also, I suggest to rephrase the clause "Consider the infinity point and work comfortably with it." in a more appropriete mathematical way... (Is "infinity" a point in the real line or in the complex plane?)
