### 2024-03-16 #140 @rpgowers
Here is my review for the article and repository of `LCD.jl`:

The paper is a succinct and readable article with a clear description and diagrams. The author outlines clearly the problem to be solved and provides useful visual diagrams showing example outputs from the package.

There are three main aspects which stand out in the article which could be improved. The first is that the example figures given do not have the arguments for `coef_num` and `coef_den` given.  Provided that this can be done concisely, giving explicit values for the arguments would help the reader understand how to use the package. Secondly, the list of references seems out of date. The preprint is referenced rather than the published journal article of the same name in "Topology and its Applications" in 2023. Finally, the authors could spend a sentence to explain to the general reader that by "indeterminations" (which I took as "indeterminacies") that this term means a zero or near-zero denominator. This is because the latter concept of small denominators will be more widely understood.

Regarding the software repository, the installation instructions as described do not work.  For example, it should read `Pkg.add(url = "https://github.com/valvarezapa/LCD.jl.git)`, although even this change is insufficient to install the package correctly. Furthermore, the list of dependencies is not given and there is no `Project.toml` file.

Once one manages a workaround for the installation, the functions provided do run and there are examples given in the source code, which is welcome. The code is also well-commented and almost all (there were a few exceptions!) the examples contained run without error and some nice-looking plots were generated. The inclusion of an introductory jupyter notebook is also appreciated. However, here too there is scope for improvement.  The main one being to implement automated tests. Second, the author should give could which generates the plots in the `Examples` folder, especially since these diagrams are the selling point of this package.

Overall this is an interesting package which solves a clear problem. Further refinements to the repository would help make it more accessible and attractive to other Julia users in the mathematical community.
