### 2024-03-25 #140 @rpgowers
> @rpgowers with the new app you should be able to do
>
> ```
> @editorialbot generate my checklist
> ```
>
> to get your checklist.

Ah great, I managed to generate a checklist and filled it in now thanks
