### 2024-04-26 #140 @valvarezapa
Hi!

Apologies for the delay. I have not started yet with this, but I find the
reviewers' comments very helpful and accurate. I am sure that their
suggestions and comments will lead to a significant improvement both in the
paper and the repository, and for that I am deeply grateful. I will start
addressing this as soon as I can (and as soon as my other
responsibilities allow me). I estimate that by the end of May I will be
able to start working on this.

Again, thank you for your patience and thanks to the reviewers.
Best,
Víctor

El jue, 25 abr 2024 a las 18:30, Luca Ferranti ***@***.***>)
escribió:

> Hi @valvarezapa <https://github.com/valvarezapa> 👋,
>
> how is it going? Have you started addressing the reviewers' comments? Any
> estimate on the timeline?
>
