### 2024-06-21 #140 @valvarezapa
Hi!

I'm really sorry for the delay. I hope to be able to address the reviewer's
comments in the following couple of weeks. I will keep you updated.

Thank you for everything.
Best,
Víctor

El lun, 17 jun 2024 a las 11:13, Luca Ferranti ***@***.***>)
escribió:

> Hi @valvarezapa <https://github.com/valvarezapa> 👋 , any updates on
> addressing the reviewers' comments?
>
