### 2024-07-01 3562de4
Manifest and Project TOML added
