+++ src/Project.toml
@@ -1,0 +1,9 @@
+authors = ["Víctor Álvarez Aparicio <valvarezaparicio@gmail.com>"]
+
+[deps]
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+Polynomials = "f27b6e38-b328-58d1-80ce-0feddd5e7a45"
+PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
+
+[compat]
+julia = "1.6.4"
