### 2024-07-01 d2f30d4
paper.tex and bib.tex updated
