+++ paper/ref.bib
@@ -10,10 +10,13 @@
 }
 
 @article{preprint,
-  title={Algorithms for computing basins of attraction associated with a rational self-map of the {H}opf fibration based on {L}yapunov exponents},
-  author={{\'A}lvarez-Aparicio, V. and García-Calcines, J.M. and Hernández-Paricio, L.J. and Rivas-Rodríguez, M.T.},
-  journal={Preprint},
-  year={2022}
+  title = {Algorithms for computing basins of attraction associated with a rational self-map of the Hopf fibration based on Lyapunov exponents},
+journal = {Topology and its Applications},
+volume = {339},
+pages = {108578},
+year = {2023},
+issn = {0166-8641},
+author = {V. Álvarez-Aparicio and J.M. García-Calcines and L.J. Hernández-Paricio and M.T. Rivas-Rodríguez},
 }
 
 @article{sage,
