+++ README.md
@@ -1,29 +1,29 @@
 ﻿# Lyapunov Cycle Detector
 
 Welcome to the *LCD* module! We present a collection of algorithms written in Julia Language which you might find very useful when approaching to certain Numerical Analysis problems involving rational maps. This kind of problems arise very often in many different areas of Mathematics and Computing like Complex Dynamics, Fractal Geometry or Algorithmic Methods. For example, when you apply Newton's method (the most currently used method to approximate the solutions of a non-linear equation) to a polynomial, the resulting function you have to iterate is a rational map.
-The basic functionality of this software is to detect attracting *n*-cycles when a complex rational map is iterated, and to plot its basins of attraction on the plane of complex numbers. In order to do this, a new method that relies in great mathematical theorems (such as Birkhoff's Ergodic Theorem and many other results given by Sullivan, recently awarded with the Abel prize) is implemented and described extensively in the *Mathematical Framework of LCD* (currently under development, since most of our results are currently on preprint. It'll be available as soon as possible).
+The basic functionality of this software is to detect attracting *n*-cycles when a complex rational map is iterated, and to plot its basins of attraction on the plane of complex numbers. In order to do this, a new method that relies in great mathematical theorems (such as Birkhoff's Ergodic Theorem and many other results given by Sullivan, recently awarded with the Abel prize) is implemented and described extensively in the article *Algorithms for computing basins of attraction associated with a rational self-map of the Hopf fibration based on Lyapunov exponents* ([https://doi.org/10.1016/j.topol.2023.108578])(https://doi.org/10.1016/j.topol.2023.108578).
 This collection of algorithms avoid many computational problems that often appears in Numerical Analysis, such as overflows caused by denominators close to zero or indeterminations that appear when the numerator and denominator are both equal to zero.
 This new method (which we will often refer to as *Lyapunov's method*) is focused on detecting attracting *n*-cycles. This approach is more general than some traditional algorithms that compute the basin of attraction of the fixed points of a given rational map. Also, another advantage of the algorithms we present is that they do not need any previous knowledge about the fixed points (nor to compute them beforehand) of the rational map in order to compute the basins of attraction.
-Some of this improvements come with a prize; this collection of algorithms might not be the fastest option to compute only the basins of attraction of fixed points of a rational map, since we want to detect not only the attracting *1*-cycles (fixed points) but also any attracting *n*-cycle. However, we benefit from some built-in Julia macros for Parallel Programming and Multithreading in order to make this algorithms work in a reasonable amount of time.
+Some of these improvements come with a prize; this collection of algorithms might not be the fastest option to compute only the basins of attraction of fixed points of a rational map, since we want to detect not only the attracting *1*-cycles (fixed points) but also any attracting *n*-cycle. However, we benefit from some built-in Julia macros for Parallel Programming and Multithreading in order to make this algorithms work in a reasonable amount of time.
 
-If you are unfamiliar with the mathematical concepts and technicallities behind these algorithms, you can check the *Mathematical Framework of LCD* and its references. If you aren't but still want to plot some cool-looking fractals or to study the basins of attractions of a rational map of your choice, you can follow the *Quick User Guide* in order to look for specific details of the usage of this package.
[…]
