+++ src/Examples.jl
@@ -1,0 +1,146 @@
+include("LyapunovCycleDetector.jl")
+
+using PyPlot
+using Images
+using Polynomials
+using BenchmarkTools
+using FileIO
+
+# IMPORTANT: LCD benefits from multi-threading, so make sure you are running Julia
+# with 4 threads! You can easily check the number of threads you are working with
+# simply by executing the next line.
+#Threads.nthreads()
+
+# Sample polynomials
+tradPol2=complex([-1.0,0.0,1.0]) # z^2-1
+tradPol3=complex([-1.0,0.0,0.0,1.0]) # z^3-1
+tradPol5=complex([-1.0,0.0,0.0,0.0,0.0,1.0]) # z^5-1
+interestingFatou=complex([-1.0,-20.0,200.0,-550.0,600.0]) # 600z^4-550z^3+200z^2-20z-1
+interestingJulia=complex([-4.0,0.0,-3.0,0.0,1.0]) # z^4-3z^2-4
+#2-cycles
+pol2cycle=complex([2.0,-2.0,0.0,1.0]) # z^3-2z+2
+anotherPol2cycle=complex([-5.29,0.0,-4.29,0.0,1.0]) # (z-im)(z+im)(z-2.3)(z+2.3)
+cubicPol2cycle=complex([-0.845-0.81*im,1.0,-0.845-0.81*im,1.0]) # (z-0.845-0.81*im)(z^2+1)
+#3-cycles
+pol3cycle=complex([-1.49175,1.0,0.0,-1.0]) # -z^3+z-1.49175
+anotherPol3cycle=complex([-1.214-0.5895*im,1.0,-1.214-0.5895*im,1.0]) # (z-(1.214+0.5895*im))(z^2+1)
+#4-cycles
+pol4cycle=complex([1.422-0.407*im,1.0,1.422-0.407*im,1.0]) # (z-(-1.422+0.407*im))(z^2+1)
+anotherPol4cycle=complex([-0.832-0.82*im,1.0,-0.832-0.82*im,1.0]) # (z-0.832-0.82*im)(z^2+1)
+
+polCycle_chebyshev=complex([1.2866*im,-1.0,-1.2866*im,1.0]) # (z-1.2866*im)(z^2-1)
+polExtra_chebyshev=complex([-2.5+1.0*im,-1.0,-1.0*(-2.5+1.0*im),1.0]) # (z-(-2.5+1.0*im))(z^2-1)
+polAux_chebyshev=complex([1.157*im,-1.0,-1.0*(1.157*im),1.0]) # (z-(1.157*im))(z^2-1)
+
+# Apply an iterative method  to the polynomial to obtain a rational function.
+# Iterative methods currently supported: Newton, Chebyshev, Halley
+pair=LCD.applyIterative_method(polAux_chebyshev,"chebyshev")
+coefficientlistnum=pair[1]
+coefficientlistden=pair[2]
+
+# Computes the list of fixed points (not necessary for the following methods)
+#fixedPoints=LCD.positionIteration_superAttractingFixedPoints(coefficientlistnum,coefficientlistden,(-1.5,1.5),(-1.5,1.5),200,5,8)[2]
+#println(fixedPoints)
+
+# Plot basins of attraction
+#=
+graphic=LCD.plotBasinsOfAttraction_Lyapunov(coefficientlistnum,coefficientlistden,(-1.5,1.5),(-1.5,1.5),200,5,8;coloringStrat="position")
+img=graphic[1]
+numColors=graphic[2]
+matrix=graphic[3]
+colors=graphic[4]
+=#
+
+# Plot ONLY the n-cycles
[…]
