+++ src/Project.toml
@@ -1,4 +1,7 @@
 authors = ["Víctor Álvarez Aparicio <valvarezaparicio@gmail.com>"]
+name = "LCD"
+uuid = "182613a9-18b6-4b88-b4d2-68357530d9e1"
+version = "1.0.0"
 
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
