### 2024-07-02 6f26350
Delete src/Project.toml
