### 2024-07-02 #140 @valvarezapa
Hi!

First, apologies for the delay. I have addressed the reviewer's comments
and updated the GitHub repository accordingly. I hope everything is correct
now. I would like to thank the reviewers again, since their generous
comments have improved both the article and the repository. I also want to
thank Luca for his patience and his work.
In the following days I'll try to include some tests, which hopefully will
ease the set up process of the package.

Thank you again for your time.
Best,
Víctor

El vie, 21 jun 2024 a las 12:21, Víctor Álvarez Aparicio (<
***@***.***>) escribió:

> Hi!
>
> I'm really sorry for the delay. I hope to be able to address the
> reviewer's comments in the following couple of weeks. I will keep you
> updated.
>
> Thank you for everything.
> Best,
> Víctor
>
> El lun, 17 jun 2024 a las 11:13, Luca Ferranti ***@***.***>)
> escribió:
>
>> Hi @valvarezapa <https://github.com/valvarezapa> 👋 , any updates on
>> addressing the reviewers' comments?
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/JuliaCon/proceedings-review/issues/140#issuecomment-2172820776>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AYNK6LH562TTWKHX7SOVQ2LZH2SE5AVCNFSM6AAAAABDGSI66OVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCNZSHAZDANZXGY>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
