### 2024-08-04 #140 @rpgowers
Apologies for taking a while to  look back at this. I have started looking at the updated repository now.
Regarding the installation instructions, and while the line:
```
add https://github.com/valvarezapa/LCD
```
does now install the package, it does not work with the following line:
```
using LCD.jl
```
This seems to be because Julia thinks the package is called `LyapunovCycleDetector`, since `using LyapunovCycleDetector` gives the error:

```
ERROR: package `LyapunovCycleDetector` did not define the expected module `LyapunovCycleDetector`, ...
```

I believe this should be fixable by simply renaming `LyapunovCycleDetector.jl` to `LCD.jl` (and making sure to rename the `include` statement in `src/Examples.jl`).

Will comment further on other aspects of the updated repository soon, this is just the most obvious (and hopefully easily fixable) one I have found.
