### 2024-08-04 #140 @rpgowers
The paper references have now been updated, so I have ticked that item now. The concepts and the example given in the paper are also now easier to understand.
