### 2024-08-10 #140 @rpgowers
The new `Examples.jl` file is appreciated and I was able to execute the functions there successfully. The Jupyter notebook is also very much appreciated.

I would ask to ensure that all the plots in the `Examples` directory have a corresponding equation in the `Examples.jl` file. I think that most of the cases do, but I believe many of the Julia set figures do not.

The readme reads better and I appreciate the updates here. One minor issue is that the urls to the papers seem to be combined into single links (e.g. [https://doi.org/10.1016/j.topol.2023.108578])(https://doi.org/10.1016/j.topol.2023.108578) ), which should be fixed.

A comment on the distinction between your method and the approach used in `DynamicalSystems.jl` (see: https://juliadynamics.github.io/DynamicalSystems.jl/previews/PR156/chaos/basins/) to find basins of attraction would also be appreciated. There is some overlap on what is trying to be achieved, however the approaches used seem to be very different. Pointing this out briefly in the paper and/or Readme would help users to better understand the strengths and best applications of the approach you present.
