### 2024-10-16 #140 @lucaferranti
Hi @valvarezapa :wave: ,

did you have time to address @rpgowers 's latest comments?
