### 2024-12-10 #140 @lucaferranti
Hi @valvarezapa :wave: ,

how is it going? Did you manage to take a look at the second round of comments from Robert?
