### 2025-01-14 #140 @valvarezapa
Apologies for the delay, and thank you Luca for your efforts and Robert for
his very useful comments to improve the repository.

I expect that by the end of this week I should have been able to address
all the issues.

Thank you all for your time,
Víctor

El jue, 9 ene 2025 a las 17:15, Luca Ferranti ***@***.***>)
escribió:

> Hi @valvarezapa <https://github.com/valvarezapa> 👋 ,
>
> how is it going with the submission? Did you manage to address Robert's
> comments? Do you have an expected timeline?
>
