### 2025-01-17 #140 @valvarezapa
Hi!

Thank you very much for your time. I think I managed to address the
reviewers comments. I renamed the .jl file so the line "using LCD.jl" works
now. I also fixed the issue with the links in the README file, and I added
a comment in this same file (at the bottom, in the "Additional comments"
section) addressing the differences between the methods in my repository
and other Julia packages as DynamicalSystems.jl. Lastly, I decided to
delete the "Julia sets" directory in the "Examples" directory, since, in
order to compute those graphics, I used a slightly modified version of the
methods in the repository. These images are a particular case of basins of
attraction of polynomials, and they can be obtained (as with any other
rational map) using the methods already contained in the repository. For
these reasons, and since this case does not even appear neither in the
Quick User Guide, nor the Examples.jl file, I consider that removing these
graphics from the "Examples" directory could be the best choice. The rest
of the graphics in the "Examples" directory have a corresponding equation
in the "Examples.jl" file.

Thank you again for your time and the reviewer for his helpful comments.
Best,
Víctor

El jue, 9 ene 2025 a las 17:15, Luca Ferranti ***@***.***>)
escribió:

> Hi @valvarezapa <https://github.com/valvarezapa> 👋 ,
>
> how is it going with the submission? Did you manage to address Robert's
> comments? Do you have an expected timeline?
>
