### 2025-03-18 #140 @rpgowers
The installation instructions still do not fully work for me, I get the following:

```
(@v1.10) pkg> add https://github.com/valvarezapa/LCD
    Updating git-repo `https://github.com/valvarezapa/LCD`
ERROR: expected the file `src/LyapunovCycleDetector.jl` to exist for package `LyapunovCycleDetector` at `/tmp/jl_QfphiW`
```
I believe that changing `name = "LyapunovCycleDetector"` to ` name = "LCD"` in `Project.toml` should fix this as everything else seems to be renamed appropriately.

I appreciate that the comparison to other packages has now been added however and the readme has been updated. The installation is the main outstanding issue for me at the moment.
